Restructure db operation files.
