+++ server/db/dbops.js
@@ -1,0 +1,31 @@
+const sqlite3 = require('sqlite3').verbose();
+const {TEST_DB_FILE_PATH, DB_FILE_PATH} = require('./config');
+const log4js = require('log4js');
+const logger = log4js.getLogger('wacai');
+
+const createDBConnection = () => {
+  console.log('process.env.DB_TEST: ', process.env.DB_TEST);
+  console.log('process.env.NODE_ENV: ', process.env.NODE_ENV);
+  let dbFilePath = process.env.DB_TEST ? TEST_DB_FILE_PATH : DB_FILE_PATH;
+  console.log('dbFilePath: ', dbFilePath);
+  let db = new sqlite3.Database(dbFilePath, sqlite3.OPEN_READWRITE, (err) => {
+    if (err) {
+      console.error(err.message);
+      logger.error(`Failed to connect to the chinook database: ${dbFilePath}`);
+      logger.error(err.message);
+    }
+    console.log('Connected to the chinook database.');
+    logger.info(`Connected to the chinook database: ${dbFilePath}`);
+  });
+
+  return db;
+};
+
+const closeDB = (db) => {
+  db.close();
+};
+
+module.exports = {
+  createDBConnection,
+  closeDB,
+};
