+++ server/router/finRouter/createFinTemplate.js
@@ -2,7 +2,8 @@ const express = require('express');
 const router = express.Router();
 const log4js = require('log4js');
 const logger = log4js.getLogger('wacai');
-const {createDBConnection, closeDB, createFinTemplate} = require('../../db/dao');
+const {createDBConnection, closeDB} = require('../../db/dbops');
+const {createFinTemplate} = require('../../db/fin/create');
 
 router.post('/createFinTemplate', (req, res) => {
   logger.info('api /createFinTemplate');
@@ -10,8 +11,11 @@ router.post('/createFinTemplate', (req, res) => {
   logger.info('create fin template data: ');
   logger.info(JSON.stringify(data));
 
+  // Get user info
+  let user = req._userInfo;
+
   let db = createDBConnection();
-  let createFinTemplatePromise = createFinTemplate(db, data);
+  let createFinTemplatePromise = createFinTemplate(db, {...data, ...user});
 
   createFinTemplatePromise.then((data) => {
     closeDB(db);
